Move ParameterReader to new package
