+++ src/main/java/uk/co/ramp/covid/simulation/RunModel.java
@@ -10,7 +10,7 @@ import com.google.gson.JsonParseException;
 import org.apache.logging.log4j.LogManager;
 import org.apache.logging.log4j.Logger;
 import uk.co.ramp.covid.simulation.parameters.CovidParameters;
-import uk.co.ramp.covid.simulation.io.ParameterReader;
+import uk.co.ramp.covid.simulation.parameters.ParameterReader;
 import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 
 import java.io.IOException;
