+++ src/test/java/uk/co/ramp/covid/simulation/ExampleParametersTest.java
@@ -5,7 +5,7 @@ import static org.junit.Assert.assertTrue;
 import org.junit.Test;
 
 import uk.co.ramp.covid.simulation.parameters.CovidParameters;
-import uk.co.ramp.covid.simulation.io.ParameterReader;
+import uk.co.ramp.covid.simulation.parameters.ParameterReader;
 import uk.co.ramp.covid.simulation.parameters.PopulationParameters;
 
 public class ExampleParametersTest {
